+++ include/nw4r/ut/ut_FileStream.h
@@ -19,7 +19,7 @@ public:
         u32 GetFileSize() const {
             return mFileSize;
         }
-        void SetFileSize(unsigned long fileSize) {
+        void SetFileSize(u32 fileSize) {
             mFileSize = fileSize;
         }
 
@@ -27,9 +27,9 @@ public:
             return mFileOffset;
         }
 
-        u32 Skip(long offset);
-        u32 Append(long offset);
-        void Seek(long offset, unsigned long origin);
+        u32 Skip(s32 offset);
+        u32 Append(s32 offset);
+        void Seek(s32 offset, u32 origin);
 
     private:
         u32 mFileSize;   // at 0x0
@@ -41,7 +41,7 @@ public:
     virtual ~FileStream() {} // at 0xC
 
     virtual u32 GetSize() const = 0;                             // at 0x40
-    virtual void Seek(long offset, unsigned long origin);        // at 0x44
+    virtual void Seek(s32 offset, u32 origin);                   // at 0x44
     virtual void Cancel();                                       // at 0x48
     virtual bool CancelAsync(AsyncCallback callback, void *arg); // at 0x4C
     virtual bool CanSeek() const = 0;                            // at 0x50
