+++ src/nw4r/ut/ut_binaryFileFormat.cpp
@@ -4,7 +4,7 @@ namespace nw4r {
 namespace ut {
 
 /* 8042a9e0 */
-bool IsValidBinaryFile(const BinaryFileHeader *header, unsigned long magic, u16 version, u16 numBlocks) {
+bool IsValidBinaryFile(const BinaryFileHeader *header, u32 magic, u16 version, u16 numBlocks) {
     if (header->magic != magic) {
         return false;
     }
